Reorganize SWF tests so they're shorter and easier to use
