+++ tests/test_swf/responses/test_activity_types.py
