+++ tests/test_swf/responses/test_domains.py
